Refactor event dispatcher: consolidate and relocate to shared libraries
