+++ services/peripherals/app/main.py
@@ -1,7 +1,6 @@
 import time
 import threading
 from src.device_monitor import NetworkMonitor
-from src.event_dispatcher import EventDispatcher
 from src.communication_interface import CommunicationInterface
 
 import logging
@@ -14,6 +13,7 @@ project_root = os.path.abspath(os.path.join(current_dir, "../../../../"))
 sys.path.insert(0, project_root)
 
 from shared_libraries.logging_config import setup_logger
+from shared_libraries.event_dispatcher import EventDispatcher
 
 def publish_heartbeat():
     while True:
